# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ChoreSync API
  version: 1.0.0
  description: >-
    Frozen Contract-First OpenAPI Specification for ChoreSync - An equitable, flexible
    household chore coordination system supporting Flatmates, Families, and Couples co-living.
servers:
  - url: http://localhost:8000
    description: Local development backend server
  - url: https://api.choresync.local
    description: Production API server
...
paths:
...
  /api/households/{household_id}:
    get:
      tags: [Households]
      summary: Get household details
      description: Retrieves a household by its unique identifier.
      operationId: getHousehold
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      responses:
        '200':
          description: Household details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/chores/{chore_id}:
    get:
      tags: [Chores]
      summary: Get chore details
      description: Retrieves complete details for a specific chore by ID.
      operationId: getChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      responses:
        '200':
          description: Chore details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chore'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Chores]
      summary: Update chore
      description: Updates fields of an existing chore definition.
      operationId: updateChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreUpdateRequest'
      responses:
        '200':
          description: Updated chore details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chore'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Chores]
      summary: Delete chore
      description: Removes a chore from the household board.
      operationId: deleteChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      responses:
        '200':
          description: Chore deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/chores/{chore_id}/nudge:
    post:
      tags: [Activity & Social]
      summary: Send gentle reminder nudge
      description: Sends a friendly reminder notification to the current assignee or open pool.
      operationId: sendNudge
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreNudgeRequest'
      responses:
        '200':
          description: Nudge sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreNudgeResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/chores/{chore_id}/comments:
    get:
      tags: [Activity & Social]
      summary: Get comments for chore
      description: Retrieves all discussion comments posted on a chore.
      operationId: getComments
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      responses:
        '200':
          description: List of comments with author information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChoreCommentWithMember'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Activity & Social]
      summary: Add comment to chore
      description: Posts a comment or question to a chore's comment feed.
      operationId: addComment
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreCommentCreateRequest'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreComment'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
  /api/uploads/photo:
    post:
      tags: [Chores]
      summary: Upload chore completion photo proof
      description: Uploads an image file to cloud object storage (S3 / Floci emulator) and returns the asset URLs.
      operationId: uploadProofPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (jpeg, png, webp, gif) up to 10MB
      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadProofPhotoResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '415':
          description: Unsupported media type
  /api/uploads/photo/{key}:
    get:
      tags: [Chores]
      summary: Retrieve photo proof from storage
      description: Streams the image binary from S3 / Floci cloud object storage.
      operationId: getProofPhoto
      parameters:
        - name: key
          in: path
          required: true
          description: The S3 object key or filename
          schema:
            type: string
      responses:
        '200':
          description: Image binary stream
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/households/{household_id}/activity:
    get:
      tags: [Activity & Social]
      summary: List household activity log
      description: Retrieves timeline of chore, swap, approval, and reward events.
      operationId: getActivities
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      responses:
        '200':
          description: List of activity logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityLog'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token encoding member_id, household_id, and role claims.
  parameters:
    HouseholdIdParam:
      name: household_id
      in: path
      required: true
      schema:
        type: string
      description: Household unique identifier
    ChoreIdParam:
      name: chore_id
      in: path
      required: true
      schema:
        type: string
      description: Chore unique identifier
...
  responses:
    BadRequestError:
      description: Bad request or validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "BAD_REQUEST"
            message: "Title is required and cannot be empty"
            code: 400
            details: {}
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "RESOURCE_NOT_FOUND"
            message: "Chore with ID c-123 was not found in household h-roommates"
            code: 404
            details: {}
...
  schemas:
    HouseholdMode:
      type: string
      enum: [flatmate, family, casual]
      description: Living arrangement mode
    MemberRole:
      type: string
      enum: [admin, member, child]
      description: Role within the household
    ChoreCategory:
      type: string
      enum: [cleaning, kitchen, yard, pets, maintenance, other]
      description: Category of the chore
    ChoreAssignmentType:
      type: string
      enum: [direct, round_robin, open_pool]
      description: Assignment strategy for the chore
    ChoreRecurrenceType:
      type: string
      enum: [none, daily, weekly, monthly, custom_days]
      description: Recurrence cadence
    ChoreRecurrenceRule:
      type: object
      properties:
        days:
          type: array
          items:
            type: string
          example: ["mon", "thu"]
    ChoreStatus:
      type: string
      enum: [unassigned, assigned, in_progress, pending_approval, completed, overdue]
      description: Current lifecycle state of the chore
...
    ActivityEventType:
      type: string
      enum:
        - chore_created
        - chore_assigned
        - chore_completed
        - approval_requested
        - chore_approved
        - chore_rejected
        - swap_proposed
        - swap_accepted
        - chore_nudge
        - reward_redeemed
        - reward_fulfilled
    HouseholdSettings:
      type: object
      required:
        - default_mode
        - allow_swaps
        - timezone
      properties:
        default_mode:
          $ref: '#/components/schemas/HouseholdMode'
        allow_swaps:
          type: boolean
        timezone:
          type: string
          example: "America/New_York"
        admin_pin:
          type: string
          description: Optional 4-digit PIN for parent/admin approval on shared devices
          example: "1234"
...
    Household:
      type: object
      required:
        - id
        - name
        - invite_code
        - settings
        - created_at
        - updated_at
      properties:
        id:
          type: string
          example: "h-roommates"
        name:
          type: string
          example: "Apartment 4B"
        invite_code:
          type: string
          example: "APT4B-SHARE"
        settings:
          $ref: '#/components/schemas/HouseholdSettings'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Member:
      type: object
      required:
        - id
        - household_id
        - name
        - role
        - points_balance
        - total_points_earned
        - streak
        - created_at
        - updated_at
      properties:
        id:
          type: string
          example: "m-sarah"
        household_id:
          type: string
          example: "h-roommates"
        name:
          type: string
          example: "Sarah Chen"
        avatar_url:
          type: string
          nullable: true
          example: "https://images.unsplash.com/photo-1534528741775-53994a69daeb"
        role:
          $ref: '#/components/schemas/MemberRole'
        points_balance:
          type: integer
          example: 145
        total_points_earned:
          type: integer
          example: 420
        streak:
          type: integer
          example: 6
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Chore:
      type: object
      required:
        - id
        - household_id
        - title
        - category
        - effort_points
        - assignment_type
        - recurrence_type
        - requires_approval
        - requires_proof
        - status
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
          example: "c-recycling-rr"
        household_id:
          type: string
          example: "h-roommates"
        title:
          type: string
          example: "Recycling & Garbage to Curb"
        description:
          type: string
          nullable: true
          example: "Separate glass, plastic, and cardboard. Roll bins to curb by 8 PM Tuesday."
        category:
          $ref: '#/components/schemas/ChoreCategory'
        effort_points:
          type: integer
          example: 20
        assignment_type:
          $ref: '#/components/schemas/ChoreAssignmentType'
        current_assignee_id:
          type: string
          nullable: true
          example: "m-liam"
        rotation_member_ids:
          type: array
          items:
            type: string
          nullable: true
          example: ["m-sarah", "m-liam", "m-maya", "m-noah"]
        current_rotation_index:
          type: integer
          nullable: true
          example: 1
        recurrence_type:
          $ref: '#/components/schemas/ChoreRecurrenceType'
        recurrence_rule:
          $ref: '#/components/schemas/ChoreRecurrenceRule'
        due_date:
          type: string
          format: date-time
          nullable: true
        requires_approval:
          type: boolean
          example: false
        requires_proof:
          type: boolean
          example: false
        status:
          $ref: '#/components/schemas/ChoreStatus'
        created_by:
          type: string
          example: "m-sarah"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ChoreUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/ChoreCategory'
        effort_points:
          type: integer
        assignment_type:
          $ref: '#/components/schemas/ChoreAssignmentType'
        current_assignee_id:
          type: string
          nullable: true
        rotation_member_ids:
          type: array
          items:
            type: string
        current_rotation_index:
          type: integer
        recurrence_type:
          $ref: '#/components/schemas/ChoreRecurrenceType'
        recurrence_rule:
          $ref: '#/components/schemas/ChoreRecurrenceRule'
        due_date:
          type: string
          format: date-time
          nullable: true
        requires_approval:
          type: boolean
        requires_proof:
          type: boolean
        status:
          $ref: '#/components/schemas/ChoreStatus'
...
    ActivityLog:
      type: object
      required:
        - id
        - household_id
        - actor_id
        - event_type
        - entity_type
        - entity_id
        - message
        - read
        - created_at
      properties:
        id:
          type: string
          example: "act-1"
        household_id:
          type: string
          example: "h-roommates"
        recipient_id:
          type: string
          nullable: true
          example: "m-liam"
        actor_id:
          type: string
          example: "m-sarah"
        event_type:
          $ref: '#/components/schemas/ActivityEventType'
        entity_type:
          type: string
          enum: [chore, swap, reward]
          example: "chore"
        entity_id:
          type: string
          example: "c-recycling-rr"
        message:
          type: string
          example: "Sarah Chen completed 'Deep Clean Kitchen Counters & Stovetop' (+25 pts)"
        read:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
    ChoreNudgeRequest:
      type: object
      required:
        - sender_member_id
      properties:
        sender_member_id:
          type: string
          example: "m-sarah"
    ChoreNudgeResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Friendly nudge from Sarah Chen to Liam Vance: 'Vacuum Living Room & Dust Shelves' is due soon!"
        email_dispatched:
          type: boolean
          example: true
        sqs_queued:
          type: boolean
          example: true
    ChoreComment:
      type: object
      required:
        - id
        - chore_id
        - member_id
        - message
        - created_at
      properties:
        id:
          type: string
          example: "comm-1"
        chore_id:
          type: string
          example: "c-recycling-rr"
        member_id:
          type: string
          example: "m-liam"
        message:
          type: string
          example: "I can take this out tomorrow morning before 8 AM."
        created_at:
          type: string
          format: date-time
    ChoreCommentWithMember:
      type: object
      required:
        - comment
      properties:
        comment:
          $ref: '#/components/schemas/ChoreComment'
        member:
          $ref: '#/components/schemas/Member'
          nullable: true
    ChoreCommentCreateRequest:
      type: object
      required:
        - member_id
        - message
      properties:
        member_id:
          type: string
          example: "m-liam"
        message:
          type: string
          example: "I will handle this tonight!"
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - code
      properties:
        error:
          type: string
          example: "RESOURCE_NOT_FOUND"
        message:
          type: string
          example: "Chore with ID c-123 was not found in household h-roommates"
        code:
          type: integer
          example: 404
        details:
          type: object
          additionalProperties: true
          example: {}
    UploadProofPhotoResponse:
      type: object
      required:
        - success
        - url
        - s3_key
        - content_type
        - size_bytes
      properties:
        success:
          type: boolean
          example: true
        url:
          type: string
          example: "/api/v1/uploads/photo/proofs/20260924-example.jpg"
        s3_key:
          type: string
          example: "proofs/20260924-example.jpg"
        content_type:
          type: string
          example: "image/jpeg"
        size_bytes:
          type: integer
          example: 10240